requirements.txt
Django>=3.1.2,<3.3.0
djangorestframework>=3.12.1,<3.13.0
flake8==3.9.2
psycopg2-binary==2.9
pyyaml==5.4.1
requests==2.26.0
django-debug-toolbar==3.2.2